Add query for boot_id.
Using uptime in testing to determine if a system has rebooted within constraints of a test is unreliable.  Systems that have rebooted occasionally report an uptime longer than that of the test determined elapsed time.
boot_id changes on every boot and can be used to determine if reboot has occurred.

Ansible was modified to use this approach: https://github.com/ansible/ansible/pull/47017